feat(agent): avatar + relative time on message role markers
Each turn's role marker now shows a small avatar and a relative timestamp instead of a bare "You"/"Assistant". The user reuses their Gravatar (UserAvatar, same as the toolbar); the agent gets the robot glyph in a matching circle. Relative time comes from the turn's first message; formatRelativeTime is extracted to a shared module so the history list and markers agree.
